pyproject.toml
[tool.poetry]
name = "mifapi"
version = "0.3.0"
description = "mifapi: Modern Image Formats (JPEG XL and AVIF) Web API"
authors = ["Evgeny Varnavskiy <varnavruz@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.2"
gunicorn = "^20.0.4"
python-multipart = "^0.0.5"
aiofiles = "^0.6.0"
APScheduler = "^3.6.3"


[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
requests = "^2.25.1"
Pillow = "^8.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
